debian-source: Add character replacement function used in name tag
By processing debian-source.bbclass, the file name is specified for the tag of
the file path in SRC_URI. If tilde (~) is included in the file name, tilde (~)
can not be used for the tag, so it fails in tag processing of the file.
This replace to underscore (_) if the tilde is included in the tag.

Signed-off-by: Nobuhiro Iwamatsu <nobuhiro.iwamatsu@miraclelinux.com>